requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
aiosqlite==0.22.1
alembic==1.18.4
amplitude-analytics==1.2.2
anaconda-anon-usage @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_3eler6mjxh/croot/anaconda-anon-usage_1710965076906/work
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
apprise==1.9.9
archspec @ file:///croot/archspec_1709217642129/work
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgi-lifespan==2.1.0
asttokens==3.0.0
async-lru==2.0.5
asyncpg==0.31.0
attrs==25.3.0
babel==2.17.0
beartype==0.22.9
beautifulsoup4==4.13.4
bleach==6.2.0
boltons @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/boltons_1699240838368/work
Brotli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_27zk0eqdh0/croot/brotli-split_1714483157007/work
cachetools==7.0.5
certifi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_70oty9s9jh/croot/certifi_1720453497032/work/certifi
cffi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_7a9c7wyorr/croot/cffi_1714483157752/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
conda @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_6fdicqiami/croot/conda_1722004613620/work
conda-content-trust @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_9fusbfzixa/croot/conda-content-trust_1714483157715/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1721662679737/work/src
conda-package-handling @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ef9phnqphe/croot/conda-package-handling_1718138279942/work
conda_package_streaming @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_6dgq200203/croot/conda-package-streaming_1718136087190/work
contourpy==1.3.2
coolname==4.1.0
cronsim==2.7
cryptography @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_31zgxm62w8/croot/cryptography_1714660690857/work
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
cyclopts==4.10.1
dataframe_image==0.2.7
dateparser==1.3.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ddkyz0575y/croot/distro_1714488254309/work
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.4
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.0
fakeredis==2.34.1
fastapi==0.115.12
fastjsonschema==2.21.1
fonttools==4.57.0
fqdn==1.5.1
frozendict @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_8b0cest_id/croot/frozendict_1713194839836/work
frozenlist==1.6.2
fsspec==2026.2.0
graphviz==0.21
greenlet==3.2.2
griffe==2.0.1
griffecli==2.0.1
griffelib==2.0.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
idna @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a12xpo84t2/croot/idna_1714398852854/work
importlib_metadata==8.7.1
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
json5==0.12.1
jsonpatch @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_3ajyoz8zoj/croot/jsonpatch_1714483362270/work
jsonpointer==2.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
libmambapy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_8b8wjifot8/croot/mamba-split_1714483537242/work/libmambapy
lupa==2.6
lxml==5.4.0
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_ad321a45td/croot/menuinst_1723567614652/work
mistune==3.1.3
more-itertools==10.7.0
MouseInfo==0.1.3
multidict==6.4.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.4
oauthlib==3.3.1
openpyxl==3.1.5
opentelemetry-api==1.40.0
orjson==3.11.7
packaging @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_81ri4yfpjw/croot/packaging_1720101866878/work
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==1.0.4
pdf2image==1.17.0
pdfkit==1.0.0
pendulum==3.2.0
pexpect==4.9.0
pillow==11.2.1
platformdirs @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/platformdirs_1701803010714/work
playwright==1.52.0
pluggy==1.6.0
prefect==3.6.23
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-key-value-aio==0.4.4
PyAutoGUI==0.9.54
pycosat @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_19qelmdbl6/croot/pycosat_1714510743067/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydocket==0.18.2
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
PyMuPDF==1.25.5
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pysocks_1699237568675/work
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
readchar==4.2.1
redis==7.3.0
referencing==0.36.2
regex==2026.2.28
requests @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_70sm12ba9w/croot/requests_1721414707360/work
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.25.1
ruamel.yaml @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/ruamel.yaml_1699241608430/work
ruamel.yaml.clib==0.2.15
rubicon-objc==0.5.0
scipy==1.15.2
semver==3.0.4
Send2Trash==1.8.3
setuptools==72.1.0
shellingham==1.5.4
six==1.17.0
slack_sdk==3.41.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.48
stack-data==0.6.3
starlette==0.46.2
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.1
tqdm @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_f76_dxtcsh/croot/tqdm_1716395948224/work
traitlets==5.14.3
truststore @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/truststore_1701805928787/work
typer==0.24.1
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uncalled-for==0.2.0
uri-template==1.3.0
urllib3 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_cao7_u9937/croot/urllib3_1718912649114/work
uvicorn==0.34.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
wheel==0.43.0
yarl==1.20.0
zipp==3.23.0
zstandard @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_054juzz3it/croot/zstandard_1714677666952/work
